+++ src/main/java/cart/dto/request/ProductRequest.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @JsonNaming(PropertyNamingStrategies.KebabCaseStrategy.class)
 public class ProductRequest {
@@ -13,11 +14,11 @@ public class ProductRequest {
     @NotEmpty
     private String name;
     @NotNull
-    private Integer price;
+    private BigDecimal price;
     @NotBlank
     private String imageUrl;
 
-    public ProductRequest(String name, Integer price, String imageUrl) {
+    public ProductRequest(String name, BigDecimal price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -31,7 +32,7 @@ public class ProductRequest {
         return imageUrl;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
