+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -3,15 +3,17 @@ package cart.dto.response;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
+import java.math.BigDecimal;
+
 @JsonNaming(PropertyNamingStrategies.KebabCaseStrategy.class)
 public class ProductResponse {
 
     private Long id;
     private String name;
-    private Integer price;
+    private BigDecimal price;
     private String imageUrl;
 
-    public ProductResponse(Long id, String name, Integer price, String imageUrl) {
+    public ProductResponse(Long id, String name, BigDecimal price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,7 +32,7 @@ public class ProductResponse {
         return imageUrl;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
